[RFE] [OVN] - IPv6 Distributed routing
Added the new NAT rule for the IPv6 GUA addresses that are allocated to
VMs. The ovn-controller running on the chassis needs this rule to start
responding GARPs (IPv4 - FIP) and neighbor advertisements (for IPv6) [1].

The chassis that has the IPv6 GUA address associated with a NAT rule
will be able to send/receive traffic without the need to centralize it
through the chassis where the router resides.

This patch implement the IPv6 DVR behaviour when enabled via [ovn]
section in ml2_conf.ini file -> enable_distributed_ipv6 = True

For VMs already created before enabling this flag, neutron will notify
in the log about the config differences when it is restarted. The NAT
rules for IPv6 GUA will be automatically applied when executing a
SYNC_MODE_REPAIR.

[1]
https://mail.openvswitch.org/pipermail/ovs-discuss/2022-December/052126.html

Closes-Bug: #1998609
Change-Id: I2ac871cc1fe2423c9481ef765ad1cf1e00354fc6
Co-authored-by: Maximilian Stinsky <maximilian.stinsky-damke@wiit.cloud>
Signed-off-by: Maximilian Stinsky <maximilian.stinsky-damke@wiit.cloud>